fix(kotlin): prevent cancellations during transactions
Updated the `transaction` function to use a NonCancellable context. As uniffi doesn't support cancellation, it was somehow causing the transaction to never complete, preventing other transactions from starting.

cherry-picked from 75217d57fe89a9d15e99d39bf02ce31efda54344
